Video Labelling API/sav-worker-master/requirements.txt
cycler==0.10.0
dask==0.8.2
decorator==4.0.9
matplotlib==1.5.1
networkx==1.11
numpy==1.11.0
pika==0.10.0
Pillow==3.2.0
protobuf==2.6.1
psycopg2==2.6.1
pyparsing==2.1.1
python-dateutil==2.5.3
python-magic==0.4.11
pytz==2016.4
scikit-image==0.12.3
scipy==0.17.0
six==1.10.0
SQLAlchemy==1.0.12
toolz==0.7.4
